pyproject.toml
[tool.poetry]
name = "ads"
version = "0.1.0"
description = ""
authors = ["Toma_san <taramis637@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
Django = "^4.0.4"
psycopg2 = "^2.9.3"
Pillow = "^9.1.1"
djangorestframework = "^3.13.1"
djangorestframework-simplejwt = "^5.2.0"
python-dateutil = "^2.8.2"
pytest-django = "^4.5.2"
pytest-factoryboy = "^2.4.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
